39 spa router

+++ app/frontend/src/views/MatchAnnouncementView.ts
@@ -1,3 +1,4 @@
+import { router } from "../routing/Router.js";
 import { deleteTournament } from "../services/tournamentService.js";
 import { Tournament } from "../Tournament.js";
 import AbstractView from "./AbstractView.js";
@@ -105,17 +106,23 @@ export default class MatchAnnouncementView extends AbstractView {
       GameType.tournament,
       this.tournament
     );
-    gameView.render();
+    router.switchView(gameView);
   }
 
   private async abortTournament() {
     try {
+      const confirmed = confirm("Do you really want to abort the tournament?");
+      if (!confirmed) return;
       await deleteTournament(this.tournament.getId());
-      const newTournamentView = new NewTournament();
-      newTournamentView.render();
+      const view = new NewTournament();
+      router.switchView(view);
     } catch (error) {
       console.error(error);
       // show error page
     }
+  }
+
+  getName(): string {
+    return "match-announcement";
   }
 }
